pyproject.toml
[project]
name = "litellm-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.75.5.post1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "litellm>=1.75.5.post1",
]
